+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [5.0.366] - 2025-06-01
+### Fixed
+- fix manifest tests
+
+## [5.0.365] - 2025-06-01
+### Fixed
+- fix manifest tests
+
 ## [5.0.364] - 2025-06-01
 ### Fixed
 - fix manifest and its tests